convergindo melhor

+++ Nosso Programa/global.h
@@ -58,6 +58,6 @@ extern int erroGrande;
 extern double z;
 
 extern int fazendoGminStepping;
-extern double gs;
+extern unsigned long long gs;
 //extern vector <double*> resultados;
 #endif
